# InMemorySagaPerister allocates less for saga with fields of simple types.