--- a/src/NServiceBus.Core.Tests/NServiceBus.Core.Tests.csproj
+++ b/src/NServiceBus.Core.Tests/NServiceBus.Core.Tests.csproj
@@ -144,7 +144,7 @@
     <Compile Include="Msmq\MsmqSubscriptionStorageTests.cs" />
     <Compile Include="Persistence\InMemory\InMemorySubscriptionStorageTests.cs" />
     <Compile Include="Persistence\InMemory\SagaWithoutUniquePropertyData.cs" />
-    <Compile Include="Persistence\InMemory\When_persisting_a_saga_with_complex_types_full_copy_is_created.cs" />
+    <Compile Include="Persistence\InMemory\When_persisting_a_saga_with_complex_types.cs" />
     <Compile Include="Persistence\InMemory\When_persisting_a_saga_with_InMemory_and_an_escalated_DTC_transaction.cs" />
     <Compile Include="Persistence\InMemory\When_persisting_a_saga_with_no_defined_unique_property.cs" />
     <Compile Include="Persistence\InMemory\When_completing_a_saga_with_no_defined_unique_property.cs" />