After this PR `InMemorySagaPersister` allocates less memory for sagas that have only fields of following types: primitive types, strings, Guids. This is provided by using [MemberwiseClone](https://msdn.microsoft.com/en-us/library/system.object.memberwiseclone(v=vs.110).aspx) instead of the JSON serializer.

This gem enables **not calling any serializer** for getting sagas that types consist of fields with following types:
- [primitive types](https://msdn.microsoft.com/en-us/library/system.type.isprimitive(v=vs.110).aspx?query=memberwiseclone): Boolean, Byte, SByte, Int16, UInt16, Int32, UInt32, Int64, UInt64, IntPtr, UIntPtr, Char, Double, and Single
- string
- Guid


### Tests
For the following saga

```
public class SagaWithGuid : ContainSagaData
{
  public Guid Guid;
}
```

initialized before tests with
```
new SagaWithGuid 
{
    Id = Guid.NewGuid(),
    Guid = Guid.NewGuid(),
    OriginalMessageId = Guid.NewGuid().ToString(),
    Originator = "Originator"
};
```

Getting saga with ` Task<TSagaData> Get<TSagaData>(Guid sagaId, ...` has following performance

Method |        Median |      StdDev | Scaled |  Gen 0 | Gen 1 | Gen 2 | Bytes Allocated/Op |
------------- |-------------- |------------ |------- |------- |------ |------ |------------------- |
Current | 6,301.7050 ns | 276.7486 ns |   1.00 | 879.00 |     - |     - |           1,669.12 |
Proposed with this PR |   399.4142 ns |  22.6731 ns |   0.06 |  91.97 |     - |     - |             162.40 |