@bording 2016-12-13
@Scooletz Do you have any before/after benchmark numbers?