@Scooletz 2016-12-14
Ping @bording @danielmarbach I provided a small benchmark.