@Scooletz 2016-12-14
@danielmarbach As you mentioned, the serializer choice is just an implementation detail. We could pick anything else. In this case I wouldn't assume having some serializer related attributes on the saga. Also, it's highly unlikely to have only simple properties and a custom serialization (or modified with attributes).