@Scooletz 2016-12-16
Ping @Particular/nservicebus-maintainers to merge as I got 2 positive reviews? :)